refactor chatbox
